+++ src/test/java/com/example/sns/post/application/PostServiceTest.java
@@ -185,51 +185,4 @@ class PostServiceTest extends ServiceTest {
         return em.createQuery("select fi from PostImage fi", PostImage.class)
                 .getResultList();
     }
-
-    @Test
-    @DisplayName("특정 게시글을 조회해야 함")
-    void findPost() throws Exception {
-        //given
-        Member author = memberRepository.save(getBasicMember());
-        Post post = postRepository.save(getBasicPost(author));
-        Member member = memberRepository.save(getBasicMember2());
-        likeService.like(member.getId(), post.getId());
-
-        //when
-        PostResponse response = postService.findPost(member.getId(), post.getId());
-
-        //then
-        assertAll(
-                () -> assertThat(response.getAuthorNickname()).isEqualTo(author.getInfo().getNickname()),
-                () -> assertThat(response.getLikeCnt()).isEqualTo(1),
-                () -> assertThat(response.getContent()).isEqualTo(post.getContent()),
-                () -> assertThat(response.getId()).isEqualTo(post.getId()),
-                () -> assertThat(response.getCreatedAt()).isEqualTo(post.getCreatedAt().toLocalDate())
-        );
-    }
-
-    @Test
-    @DisplayName("유저가 없는 경우 조회를 하면 예외가 발생해야 함")
-    void findPost_memberNotFound() throws Exception {
-        //given
-        Long notExistId = 999L;
-        Member author = memberRepository.save(getBasicMember());
-        Post post = postRepository.save(getBasicPost(author));
-
-        //when
-        assertThatThrownBy(() -> postService.findPost(notExistId, post.getId()))
-                .isInstanceOf(MemberNotFoundException.class);
-    }
-
-    @Test
-    @DisplayName("게시글이 없는 경우 조회를 하면 예외가 발생해야 함")
-    void findPost_postNotFound() throws Exception {
-        //given
-        Long notExistId = 999L;
-        Member author = memberRepository.save(getBasicMember());
-
-        //when
-        assertThatThrownBy(() -> postService.findPost(author.getId(), notExistId))
-                .isInstanceOf(PostNotFoundException.class);
-    }
 }
